pyproject.toml
[project]
name = "stvorec-02"
version = "0.1.0"
description = ""
authors = [
    {name = "Hamper",email = "barca.teo@gmail.com"}
]
readme = "README.md"
requires-python = "~3.11"
dependencies = [
    "langchain (>=1.1.0,<2.0.0)",
    "langgraph (>=1.0.4,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "langchain-openai (>=1.1.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[dependency-groups]
dev = []


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
